backend/uets_rm/requirements.txt
djangorestframework == 3.11.0
pandas
numpy
joblib
sklearn
psycopg2